requirements.txt
numpy
dimod
dwave-ocean-sdk
